[PSEH3] Fix _SEH3_LEAVE
Previously, when leave was used in the except or finally block of a nested try block, it would jump back to the start of the except/finally block, resulting in an endless loop. This is fixxed by jumping back to a label at the beginning of the try block, which is only visible from within the try block itself and from there to the end of the seh block.
Discovered by MS SEH tests.