[PSEH3] Fix unlinking of EH frames during unwind.
On Windows the function that calls the language specific exception handler function registers it's own exception handler, so the top of the exception chain does not point to our own handler. Therefore we need to take that into account when unwinding the stack and removing each handler as we go.