[STUBS] On x64 do not raise exception in __wine_spec_unimplemented_stub
On x86 the stdcall calling convention requires that the callee cleans up the stack. Stub functions with the wrong number of parameters would therefore break the stack. This is why stub functions without parameter annotations throw an exception.
On x64 there is only one calling convention, where the caller is responsible for cleaning up the stack, so we can safely implement a stub function with the wrong number of parameters without breaking the stack.